content/base/test/chrome/test_bug635835.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=635835
-->
<window title="Mozilla Bug 635835"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=635835"
     target="_blank">Mozilla Bug 635835</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
SimpleTest.waitForExplicitFinish();
const SHOW_ALL = Components.interfaces.nsIDOMNodeFilter.SHOW_ALL;

addLoadEvent(function() {
  var walker = document.createTreeWalker(document, SHOW_ALL, null, true);
  try {
    walker.currentNode = {};
    walker.nextNode();
  }
  catch (e) {
    // do nothing - this is a crash test
  }
  ok(true, "Crash test passed");
  SimpleTest.finish();
});
  ]]></script>
</window>
